requirements.txt
faker==24.11.0  
kafka-python==2.0.2  
dash==2.16.1  
plotly==5.21.0  
pandas==2.2.2  
sqlalchemy==2.0.29  
psycopg2-binary==2.9.9  # For PostgreSQL (adjust if using another DB)